+++ alh_pdf_view_ios/ios/Classes/AlhPdfViewPlugin.m
@@ -1,8 +1,0 @@
-//
-//  AlhPdfViewPlugin.m
-//  alh_pdf_view_ios
-//
-//  Created by André Börger on 26.03.24.
-//
-
-#import <Foundation/Foundation.h>