+++ alh_pdf_view_ios/ios/alh_pdf_view_ios/Sources/alh_pdf_view_ios/UIntExtension.swift
@@ -5,7 +5,7 @@
 //  Created by André Börger on 29.03.22.
 //
 
-import Foundation
+import UIKit
 
 extension UInt {
     func toUIColor() -> UIColor {